Move analysis tasks to measurement.
