+++ python/lsst/faro/measurement/PatchMeasurement.py
