+++ python/lsst/faro/measurement/PatchMetricMeasureTasks.py
