+++ python/lsst/faro/measurement/TractMeasurement.py
