+++ python/lsst/faro/measurement/VisitMeasurement.py
