+++ python/lsst/faro/measurement/__init__.py
@@ -1,6 +1,10 @@
 from .example_VisitMeasureTasks import *
 from .GeneralMeasureTasks import *
-from .MatchedCatalogMeasureTasks import *
-from .PatchMeasureTasks import *
-from .VisitMeasureTasks import *
-from .TractMeasureTasks import *
+from .MatchedCatalogMetricMeasureTasks import *
+from .PatchMetricMeasureTasks import *
+from .VisitMetricMeasureTasks import *
+from .TractMetricMeasureTasks import *
+from .MatchedCatalogMeasurement.py import *
+from .VisitMeasurement.py import *
+from .TractMeasurement.py import *
+from .PatchMeasurement.py import *
