+++ python/lsst/faro/preparation/__init__.py
@@ -1,5 +1,1 @@
-from .MatchedCatalogs import *
-from .MatchedCatalogsAnalysis import *
-from .PatchAnalysis import *
-from .TractAnalysis import *
-from .VisitAnalysis import *
+from .MatchedPrepTasks import *
